+++ Spooksoft.Xml.Serialization.Test/Models/Binary/BinaryModel.cs
@@ -1,0 +1,15 @@
+﻿using Spooksoft.Xml.Serialization.Attributes;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Spooksoft.Xml.Serialization.Test.Models.Binary
+{
+    public class BinaryModel
+    {
+        [XmlBinary]
+        public byte[]? BinaryData { get; set; }
+    }
+}