+++ Spooksoft.Xml.Serialization.Test/Models/Map/DerivedValue1.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Spooksoft.Xml.Serialization.Test.Models.Map
+{
+    public class DerivedValue1 : BaseValue
+    {
+        public int IntValue { get; set; }
+    }
+}