+++ Changelog.md
@@ -5,3 +5,5 @@
 # 1.0.1
 
 * [new] Added support for single-dimensional arrays
+* [new] Added support for maps (e.g. Dictionary<,>)
+* [new] Added support for binary serialization (only for byte[] properties)